// src/types/mock-data.type.ts
import { Genre } from "./genre.type";

export type MockData = {
  titles: string[];
  descriptions: string[];
  genres: Genre[];
  videoPreviewLinks: string[];
  videoLinks: string[];
  actors: string[];
  directors: string[];
  users: string[];
  emails: string[];
  avatarPaths: string[];
  passwords: string[];
  posterPaths: string[];
  backgroundPicturePaths: string[];
  backgroundColors: string[];
};